Update advanced-logger.js in order to comply with Winston configuration
It seems that you have an error in the configuration of the logger. As a matter of fact, if a configure the logger to have a max size of 1mb you should configure the winston.logger maxsize variable in this way filesize = config.maxsize * 1048576 and not filesize = config.maxsize * 1073741824 (since it is 1000 Mb).